requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
blis==1.3.3
cachetools==6.2.2
catalogue==2.0.10
catboost==1.2.8
certifi==2025.11.12
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
cymem==2.0.13
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
duckduckgo_search==8.1.1
easyocr==1.7.2
executing==2.2.1
fastapi==0.128.0
fastjsonschema==2.21.2
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.61.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.7.0
gensim==4.4.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.46.0
google-auth-httplib2==0.2.1
google-genai==1.56.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==3.2.0
h5py==3.15.1
hf-xet==1.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.3.5
hyperframe==5.2.0
idna==2.10
ImageHash==4.3.2
ImageIO==2.37.2
instaloader==4.15
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keras==3.12.0
kiwisolver==1.4.9
langdetect==1.0.9
lark==1.3.1
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.45.1
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.15
namex==0.1.0
narwhals==2.12.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6
ninja==1.13.0
notebook==7.5.0
notebook_shim==0.2.4
numba==0.62.1
numpy==2.3.5
openai==0.28.0
openai-whisper==20250625
opencv-python-headless==4.13.0.90
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.0.0
platformdirs==4.5.0
playwright==1.56.0
plotly==6.5.0
praw==7.8.1
prawcore==2.4.0
preshed==3.0.12
primp==0.15.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
proxyproviders==0.2.1
psutil==7.1.3
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydub @ https://github.com/jiaaro/pydub/archive/refs/heads/master.zip#sha256=6cf60793cf78fc23820049d3241550dfe5df5968b28f16abae801b5680d06d33
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.5
PyPDF2==3.0.1
pypdfium2==5.3.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytz==2025.2
PyWavelets==1.9.0
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
stack-data==0.6.3
starlette==0.50.0
streamlit==1.53.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
terminado==0.18.1
text-unidecode==1.3
thinc==8.3.10
threadpoolctl==3.6.0
tifffile==2026.1.28
TikTokApi==7.2.1
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==5.0.0
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ucimlrepo==0.0.7
update-checker==0.18.0
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.40.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.14
weasel==0.4.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
whisper==1.1.10
widgetsnbextension==4.0.15
wrapt==2.0.1
xgboost==3.1.2
yarl==1.22.0
youtube-transcript-api==1.2.3
yt-dlp==2025.12.8
